refactor(web): isolate session tree DOM adapter
